refactor: move Chart to global folder
