+++ src/global/components/chart/graphStyle.js
