+++ src/global/components/chart/utils/graph-utils.js
