+++ src/global/components/chart/utils/typing-utils.js
